% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_map.R, R/build_table.R
\name{build_table}
\alias{build_table}
\title{build_map}
\usage{
build_table(fmt, ...)

build_table(fmt, ...)
}
\arguments{
\item{fmt}{the package to use to format the dataset}

\item{...}{all the necessary values to \code{extract_data}}

\item{tls}{the background tiles for the map}
}
\value{
a leaflet map

a table in one of the available formats
}
\description{
Format a dataset
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
