% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCG}
\alias{CCG}
\title{CCG}
\format{
A \code{sf} dataframe with only one \code{CCG} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CCG)}.
}
\usage{
CCG
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg} 4326) for all 113 \emph{Clinical Commissioning Groups} in England and Wales
(106 \emph{Sub-ICB Location} in England and 7 \emph{Local Health Boards} in Wales).
}
\details{
Built by dissolving the \code{LSOA} boundaries using the \code{lookups} table.
}
\keyword{datasets}
