% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCGo}
\alias{CCGo}
\title{CCGo}
\format{
A \code{sf} dataframe with only one \code{CCG} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CCG)}.
}
\usage{
CCGo
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 113 \emph{Clinical Commissioning Groups} in England and Wales
(106 \emph{Sub-ICB Location} in England and 7 \emph{Local Health Boards} in Wales).
}
\details{
This is the result of the conversion of the original ONS boundaries in \emph{shapefile} format
with a \emph{OSGB36} CRS (\emph{EPSG} 27700), simplified at a \emph{20\%} rate.

Last Updated: Apr-22
}
\keyword{datasets}
