% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTRY}
\alias{CTRY}
\title{CTRY}
\format{
A \code{sf} dataframe with only one \code{CTRY} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_CTRY)}.
}
\usage{
CTRY
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for England, plus Wales.
}
\details{
Built by dissolving the \code{RGN} boundaries using the \code{lookups} table.
}
\keyword{datasets}
