% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{domains}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The identifier for the Domain }
\item{\code{ name }}{ A short name for the Domain, reference for the \code{dt_} datasets }
\item{\code{ description }}{ A possibly longer description for the Domain }
\item{\code{ ordering }}{ The order for presentation }
\item{\code{ loaded }}{ Indicates if the domain has some data actually loaded into the package }
}
}
\usage{
domains
}
\description{
The list of tables with some of their characteristics
}
\keyword{datasets}
