% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_ethnicity}
\alias{dt_ethnicity}
\title{Ethnicities and Identities}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS021 }}{ Ethnic group }
\item{\code{ TS022 }}{ Ethnic group (detailed) }
\item{\code{ TS023 }}{ Multiple ethnic group }
\item{\code{ TS027 }}{ National identity - UK }
\item{\code{ TS028 }}{ National identity (detailed) }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_ethnicity
}
\description{
Ethnicities and Identities
}
\keyword{datasets}
