% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_gender}
\alias{dt_gender}
\title{Gender}
\format{
A data.table in \emph{long} format with the following columns:
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}

The dataset includes the following Census Tables:
\describe{
\item{\code{ TS077 }}{ Sexual orientation }
\item{\code{ TS079 }}{ Sexual orientation (detailed) }
\item{\code{ TS078 }}{ Gender identity }
\item{\code{ TS070 }}{ Gender identity (detailed) }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_gender
}
\description{
Resident Population by Sexual Orientation and Gender Identity
}
\keyword{datasets}
