% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_geography}
\alias{dt_geography}
\title{Geographic Properties}
\format{
A data.table in \emph{long} format with the following columns
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}
}
\usage{
dt_geography
}
\description{
Some characteristics about Zones, such as Area, Perimeter, Centroid, ...
}
\details{
This table is \emph{not} part of the ONS Census releases.
}
\keyword{datasets}
