% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_health}
\alias{dt_health}
\title{Health and Care}
\format{
A data.table in \emph{long} format with the following columns:
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}

The dataset includes the following Census Tables:
\describe{
\item{\code{ TS037 }}{ General health }
\item{\code{ TS038 }}{ Disability  }
\item{\code{ TS039 }}{ Provision of unpaid care }
\item{\code{ TS040 }}{ Number of disabled people in household }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_health
}
\description{
Resident Population by General Health, Disability, and Provision of Unpaid Care,
plus Household by Number of Disabled Persons
}
\keyword{datasets}
