% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_housing}
\alias{dt_housing}
\title{Housing}
\format{
A data.table in \emph{long} format with the following columns:
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}

The dataset includes the following Census Tables:
\describe{
\item{\code{ TS044 }}{ Accommodation Type }
\item{\code{ TS054 }}{ Tenure }
\item{\code{ TS051 }}{ Number of Rooms }
\item{\code{ TS053 }}{ Occupancy Rating for Rooms }
\item{\code{ TS050 }}{ Number of Bedrooms }
\item{\code{ TS052 }}{ Occupancy Rating for Bedrooms }
\item{\code{ TS046 }}{ Central Heating }
\item{\code{ TS045 }}{ Car or Van availability }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_housing
}
\description{
Households by Accommodation type, Tenure of Household, Number of Rooms and Bedrooms (total count and Occupancy Rating),
Type of Central Heating, Number of Cars or Vans.
}
\keyword{datasets}
