% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_language}
\alias{dt_language}
\title{Languages and UK Proficiency}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS024 }}{ Main language (detailed) }
\item{\code{ TS025 }}{ Household language }
\item{\code{ TS029 }}{ Proficiency in English }
\item{\code{ TS026 }}{ Multiple main languages in households }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_language
}
\description{
Languages and UK Proficiency
}
\keyword{datasets}
