% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_metrics}
\alias{dt_metrics}
\title{dt_metrics}
\format{
A data.table in \emph{long} format with the following columns:
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Metric, foreign key to the \code{id} column in the \code{metrics} table }
\item{\code{ value }}{ The value for the Metric in the Zone}
}
}
\usage{
dt_metrics
}
\description{
Dataset containing the values of the \emph{metrics} built using the variables and the operations listed in the \code{metrics} table.
By design, and differently from the other \verb{dt_*} tables (besides the fact that it's \emph{not} part of the ONS Census releases),
the values are given for \emph{all} the geographies as you cannot aggregate values from lower to higher levels of the geographic hierarchy.
}
\keyword{datasets}
