% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_migration}
\alias{dt_migration}
\title{Migration}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS004 }}{ Country of birth }
\item{\code{ TS012 }}{ Country of birth (detailed) }
\item{\code{ TS005 }}{ Passports held }
\item{\code{ TS013 }}{ Passports held (detailed) }
\item{\code{ TS015 }}{ Year of arrival in UK }
\item{\code{ TS018 }}{ Age of arrival in the UK }
\item{\code{ TS016 }}{ Length of residence }
\item{\code{ TS019 }}{ Migrant Indicator }
\item{\code{ TS020 }}{ Number of non-UK short-term residents by sex }
\item{\code{ TS041 }}{ Number of Households }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_migration
}
\description{
Migration
}
\keyword{datasets}
