% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_population}
\alias{dt_population}
\title{Population and Households}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS001 }}{ Number of usual residents }
\item{\code{ TS002 }}{ Legal partnership status }
\item{\code{ TS006 }}{ Population density }
\item{\code{ TS003 }}{ Household composition }
\item{\code{ TS017 }}{ Household size }
\item{\code{ TS010 }}{ Living arrangements }
\item{\code{ TS011 }}{ Households by deprivation dimensions }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_population
}
\description{
Population and Households
}
\keyword{datasets}
