% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_veterans}
\alias{dt_veterans}
\title{UK Armed Forces Veterans}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS071 }}{ Previously served in the UK armed forces }
\item{\code{ TS072 }}{ Number of people in household who have previously served in UK armed forces }
\item{\code{ TS073 }}{ Population who have previously served in UK armed forces in communal establishments and in households }
\item{\code{ TS074 }}{ Household Reference Person indicator of previous service in UK armed forces }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_veterans
}
\description{
UK Armed Forces Veterans
}
\keyword{datasets}
