% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_work}
\alias{dt_work}
\title{Work and Commuting}
\format{
A data.table in \emph{long} format with the following columns:
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}

The dataset includes the following Census Tables:
\describe{
\item{\code{ TS066 }}{ Economic Activity Status }
\item{\code{ TS063 }}{ Occupation }
\item{\code{ TS064 }}{ Occupation - minor groups }
\item{\code{ TS062 }}{ NS-SeC }
\item{\code{ TS060 }}{ Industry }
\item{\code{ TS059 }}{ Hours worked }
\item{\code{ TS065 }}{ Unemployment history }
\item{\code{ TS058 }}{ Distance travelled to Work }
\item{\code{ TS061 }}{ Method of Travel to Work }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_work
}
\description{
Resident Population by Economic Activity Status, Occupation, Industry, National Statistics Socio-economic Classification (NS-SeC),
Hours worked, Unemployment history, Method of Travel to Work and Distance travelled.
}
\keyword{datasets}
