% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{extract_data}
\usage{
extract_data(tp)
}
\arguments{
\item{tp}{the acronym of a geographic hierarchy}

\item{x}{the id of a variable}

\item{cd}{the code of a zone}
}
\value{
a data.table
}
\description{
Build a dataset for specified variable and zone or (part of) a geographic hierachy,
eventually adding a percentage vs reference data and/or upper level hierarchy
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
