% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{localities}
\alias{localities}
\title{localities}
\format{
A data.table with the following columns:
\describe{
\item{\code{ LOC }}{ The unique ONS code for the Location }
\item{\code{ LOCd }}{ The description for the Location }
\item{\code{ LOCo }}{ A description suitable for ordering }
\item{\code{ x_lon}}{ The longitude coordinate of the Location }
\item{\code{ y_lat}}{ The latitude coordinate of the Location }
\item{\code{ OA }}{ The ONS code for the Output Areas that includes the Location}
}

For further details, see the \emph{Index of Place Names in Great Britain} page of the
\href{https://geoportal.statistics.gov.uk/datasets/index-of-place-names-in-great-britain-december-2022/about}{ONS Open Geography portal}.
}
\usage{
localities
}
\description{
The list of all known Locality in England and Wales, and the corresponding \emph{Output Areas} they're in.
}
\details{
The table contains \emph{all} UK geographical postcodes, both \emph{live} and \emph{terminated} (2,334,674, as of NOV-22).

Last Updated: Dec-21
}
\keyword{datasets}
