% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookups}
\alias{lookups}
\title{Lookups}
\format{
A data.table with the following columns:
\describe{
\item{\code{OA}}{Output Area. Total units: . Base Layer. Direct Parent: LSOA }
\item{\code{LSOA}}{Lower Layer Super Output Area. Total units: . Direct Parent: MSOA }
\item{\code{MSOA}}{Middle Layer Super Output Area. Total units: 9,370. Direct Parent: LTLA }
\item{\code{LTLA}}{Lower Tier Local Authority. Total units: 363. Direct Parent: UTLA }
\item{\code{UTLA}}{Upper Tier Local Authority. Total units: 93. Last updated: May-21. Built From LAD. Direct Parent: RGN }
\item{\code{RGN}}{Region. Total units: 10. Direct Parent: CTRY }
\item{\code{CTRY}}{Country. Total units: 2. Direct Parent: NAT }
\item{\code{PCON}}{Westminster Parliamentary Constituency. Total units: 650. Direct Parent: RGN }
\item{\code{WARD}}{Electoral Ward. Total units: 8,875. Direct Parent: LTLA }
\item{\code{PAR}}{Civil Parish or Unparished; Total units: 12,415. Direct Parent: UTLA }
\item{\code{CCG}}{Clinical Commissioning Group. Total units: 106. Direct Parent: CTRY }
}

For further details, see the \code{Lookups} section within the \href{https://geoportal.statistics.gov.uk/}{ONS Open Geography portal}
}
\usage{
lookups
}
\description{
This dataset contains both a complete list of the \emph{Output Areas}, the smallest statistical geographic area in both England and Wales,
and a mapping between them and all the other Geographies contained in the package, listed in the table \code{zone_types}.
}
\keyword{datasets}
