% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSOA}
\alias{LSOA}
\title{LSOA}
\format{
A \code{sf} dataframe with only one \code{LSOA} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_LSOA)}.
}
\usage{
LSOA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg} 4326) for all 35,672 \emph{Lower Layer Super Output Areas} in England and Wales, \strong{Census 2021}.
}
\details{
Built by dissolving the \code{OA} boundaries using the \code{lookups} table.
}
\keyword{datasets}
