% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{main_refs}
\alias{main_refs}
\title{Main References}
\format{
A data.table in \emph{long} format with the following columns
\describe{
\item{\code{ sub_id }}{ The id for the reference variable in a table }
\item{\code{ main_id }}{ The actual id for the variable in the dataset to refer to  }
\item{\code{ name }}{ A name for the variable, usually shorter than the official one }
\item{\code{ total }}{ The reference Total. There could be different \emph{similar} totals among the various tables, the percentage won't exactly sum to 100.) }
\item{\code{ diff }}{  The difference between the true total of the reference variable in a table and its substitution }
}
}
\usage{
main_refs
}
\description{
A convenience table to group all and only the main \emph{reference} Variables, with the substituted codes and the differences in units.
}
\keyword{datasets}
