% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_missing.R
\name{map_missing}
\alias{map_missing}
\title{map_missing}
\usage{
map_missing(tp)
}
\arguments{
\item{tp}{the acronym of the area: one in \code{PCON}, \code{WARD}, \code{PAR}, \code{CCG}}
}
\value{
a leaflet map
}
\description{
Build a leaflet map for the overlapping PAR or WARD locations
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
