% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_neighbours.R
\name{map_neighbours}
\alias{map_neighbours}
\title{map_neighbours}
\usage{
map_neighbours(x, mt = 1)
}
\arguments{
\item{x}{the ONS code of an area}

\item{tp}{the correct type of the requested area, in case the code \code{x} refers to multiple types}
}
\value{
a leaflet map
}
\description{
Build a leaflet map with an area with its neighbours
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
