% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nonexact.R
\name{map_nonexact}
\alias{map_nonexact}
\title{map_nonexact}
\usage{
map_nonexact(tp)
}
\arguments{
\item{tp}{the acronym of the area: one in \code{PCON}, \code{WARD}, \code{PAR}, \code{CCG}}
}
\value{
a leaflet map
}
\description{
Build a leaflet map for one of the four hierarchies that has no exact lin
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
