% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{metrics}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The id for the metric }
\item{\code{ name }}{ The name for the metric }
\item{\code{ var1_id }}{ The id for the \emph{first} variable }
\item{\code{ type1 }}{ The \emph{type} of the \emph{first} variable: \code{V}ariable or \code{M}etric }
\item{\code{ var2_id }}{ The id for the \emph{second} variable }
\item{\code{ type2 }}{ The \emph{type} of the \emph{second} variable: \code{V}ariable or \code{M}etric }
\item{\code{ ops }}{ The operation to be applied }
}
}
\usage{
metrics
}
\description{
A listing of binary operations on how to build basic metrics.
Once actioned using the \code{build_metrics} function, the values are stored for every possible level of geographic hierarchy in the \code{dt_metrics} table.
}
\keyword{datasets}
