% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSOA}
\alias{MSOA}
\title{MSOA}
\format{
A \code{sf} dataframe with only one \code{MSOA} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_MSOA)}.
}
\usage{
MSOA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 7,264 \emph{Medium Layer Super Output Areas} in England and Wales, \strong{Census 2021}.
}
\details{
Built by dissolving the \code{LSOA} boundaries using the \code{lookups} table.
}
\keyword{datasets}
