% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA}
\alias{OA}
\title{OA}
\format{
A \code{sf} dataframe with only one \code{OA} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_OA)}.
}
\usage{
OA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 188,880 \emph{Output Areas} in England and Wales, \strong{Census 2021}.
}
\keyword{datasets}
