% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCON}
\alias{PCON}
\title{PCON}
\format{
A \code{sf} dataframe with only one \code{PCON} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_PCON)}.
}
\usage{
PCON
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 573 \emph{Westminster Parliamentary Constituency} in England and Wales.
}
\details{
Built by dissolving the \code{LSOA} boundaries using the \code{lookups} table.

Last Updated: Dec-21
}
\keyword{datasets}
