% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCONo}
\alias{PCONo}
\title{PCONo}
\format{
A \code{sf} dataframe with only one \code{PCON} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_PCON)}.
}
\usage{
PCONo
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg} 4326) for all 573 \emph{Westminster Parliamentary Constituency} in England and Wales.
}
\details{
This is the result of the conversion of the original ONS boundaries in \emph{shapefile} format
with a \emph{OSGB36} CRS (\emph{epsg} 27700), simplified at a \emph{20\%} rate.
}
\keyword{datasets}
