% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcodes}
\alias{postcodes}
\title{postcodes}
\format{
A data.table with the following columns:
\describe{
\item{\code{ PCU }}{ A postcode Unit, expressed in 7-chars }
\item{\code{ OA }}{ The ONS code for an Output Areas }
\item{\code{ PCS }}{ A postcode Sector }
}

For further details on UK postcodes, see the \emph{ONS Postcodes} tagged pages of the
\href{https://geoportal.statistics.gov.uk/search?collection=Dataset&q=postcodes&sort=-created&tags=onspd}{ONS Open Geography portal}.
}
\usage{
postcodes
}
\description{
A lookup table between postcode \emph{Units} (\code{PCU}) and \emph{Output Areas}.
}
\details{
The table contains \emph{all} UK geographical postcodes, both \emph{live} and \emph{terminated} (2,334,674, as of NOV-22).

This table is part of my \href{https://github.com/lvalnegri/RpostcodesUK}{RpostcodesUK $R$ package}.
It's been included here only for search purposes, and it's not part of any ONS Census releases.
}
\keyword{datasets}
