% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdata.R
\name{qdata}
\alias{qdata}
\title{Extract Census data}
\usage{
qdata(
  idv,
  zns = NULL,
  tpz = NULL,
  smz = FALSE,
  spz = NULL,
  pct = FALSE,
  rfv = NULL,
  verbose = FALSE
)
}
\arguments{
\item{idv}{the id of a Variable (see the column \code{id} in the \code{vars} table)}

\item{zns}{the ONS code(s) of one or multiple \emph{similar} Zone(s) (see the column \code{id} in the \code{zones} table).
To check for the Zone code given its name use the function \code{qznames}.}

\item{tpz}{the acronym of a geographic hierarchy (see the column \code{type} in the \code{zone_types} table)}

\item{smz}{if \code{TRUE} and multiple Zones are requested, returns the aggregation instead of each Zones values}

\item{spz}{the ONS code of a Zone of an upper level geographic hierarchy to filter out results
(use the `` function or
directly see the \code{lookups} table from left to right using the `level` column in the `zone_types` table)}

\item{pct}{if \code{TRUE}}

\item{rfv}{the \emph{id} of a reference Variable (see the column \code{var_id} in the \code{vars_refs} table)
to create rates and/or indices}

\item{verbose}{if \code{TRUE} print all messages and warnings.}
}
\value{
a data.table
}
\description{
Build a dataset for a specified Variable and either one Zone or (part of) a Geographic Hierarchy
(see the table \code{zone_types}), eventually adding a percentage vs a \emph{reference} variable and/or upper level hierarchy
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
