% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcu.R
\name{qpcu}
\alias{qpcu}
\title{Query about a Postcode Unit}
\usage{
qpcu(pcu, ...)
}
\arguments{
\item{pcu}{a string possibly representing a UK postcode unit}

\item{...}{the remaining arguments for the \code{qzones} function}
}
\value{
a character vector, a \code{data.table} or an \code{sf} polygons object, depending on the values provided for the arguments in \code{...}
}
\description{
Extract the Output Area that contains a Postcode Unit, or all the Zones included in a parent Zone
}
\details{
Build a dataset or digital boundaries for one or more Zone(s) or (part of) a Geographic Hierarchy (see the table \code{zone_types}).
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
