% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzones.R
\name{qznames}
\alias{qznames}
\title{Query about the codes attached to the name of a Geographic Zone}
\usage{
qznames(x, tpe = NULL, no_cases = FALSE, exact = FALSE)
}
\arguments{
\item{x}{a string representing a pattern for one or more potential Zone name(s)}

\item{tpe}{a string indicating a hierarchy type to be filtered out (see the table \code{zone_types} for more details)}

\item{no_cases}{logical, override cases in \code{x}}

\item{exact}{logical, search only for the provided string}
}
\value{
a string indicating a code if the result is a single zone,
or a \code{data.table} with codes, names and geographic hierarchy type, if multiple zones have been found
}
\description{
Extract all the codes for the Zones that match the given pattern, and an optional hierarchy.
If there is only one Zone, only its code is returned as a vector.
Otherwise, the output is a data.table with matched names and correspondet codes and hierarchy type.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
