% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qzones.R
\name{qzones}
\alias{qzones}
\title{Query about one or multiple Geographic Zone(s)}
\usage{
qzones(zns, hrz = NULL, nms = FALSE, tbl = TRUE)
}
\arguments{
\item{zns}{the ONS code(s) of one or more Zones of the same type, unless asking only for the names (see the column \code{id} in the \code{zones} table)}

\item{hrz}{the type of output:
- \code{is.null(hrz)}: returns only the name(s) of the \code{zns} Zone(s), overriding all other options
- \code{hrz == "N"}: returns the \emph{neighbours} of all the provided \code{zns} Zones
- \code{level(hrz) < level(zns)}: returns all the smaller \code{hrz} Zones included in \code{zns}
- \code{level(hrz) > level(zns)}: returns all the Zones of the same type of \code{zns}
included in all the bigger \code{hrz} Zone(s) that contains \code{zns}}

\item{nms}{if \code{TRUE}, includes also a column with the name of the Zone(s)}

\item{tbl}{if \code{TRUE} exports a \code{data.table}; otherwise, the correspondent digital boundaries in \code{sf} format}
}
\value{
a character vector, a \code{data.table}, or an \code{sf} polygons object, depending on the values provided for the arguments
}
\description{
Extract the names of the provided zones or all the Zones included in the same parent Zone
}
\details{
Build a dataset or digital boundaries for one or more Zone(s) or (part of) a Geographic Hierarchy (see the table \code{zone_types}).
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
