% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summaries}
\alias{summaries}
\title{Summaries}
\format{
A data.table with the following columns
\describe{
\item{\code{ table_code }}{ The code for the variables, foreign key to the \code{code} column in the \code{tables} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ ons_name }}{ The name for the Variable as included in the original ONS files }
\item{\code{ E }}{ The total value for the Variable at \emph{England} level }
\item{\code{ W }}{ The total value for the Variable at \emph{Wales} level }
\item{\code{ T }}{ The total value for the Variable at \emph{Nation} (England and Wales) level }
}
}
\usage{
summaries
}
\description{
The Totals for all Variables at \emph{Countries} and \emph{Nation} levels.
The table also includes the original ONS names for the variables as stated in the deployed files.
}
\keyword{datasets}
