% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\docType{data}
\name{tiles.lst}
\alias{tiles.lst}
\title{tiles.lst}
\format{
An object of class \code{list} of length 22.
}
\usage{
tiles.lst
}
\description{
List of background tiles for \code{leaflet} maps
}
\references{
\href{https://leaflet-extras.github.io/leaflet-providers/preview/}{Leaflet-providers preview}.

To work correctly, needs to be paired with \code{add_maptile}.
}
\keyword{datasets}
