% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UTLA}
\alias{UTLA}
\title{UTLA}
\format{
A \code{sf} dataframe with only one \code{UTLA} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_UTLA)}.
}
\usage{
UTLA
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{epsg} 4326) for all 174 \emph{Upper Tier Local Authorities} in England and Wales.
}
\details{
Built by dissolving the \code{LTLA} boundaries using the \code{lookups} table.
}
\keyword{datasets}
