% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars}
\alias{vars}
\title{vars}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The code for the variables  }
\item{\code{ ordering }}{ The order to use for reporting purposes  }
\item{\code{ description }}{ The description for the variables }
\item{\code{ table_id }}{ The Domain the table is included, a foreign reference to the column \code{id} in the \code{domains} table }
\item{\code{ ref_id }}{ The id for the variable to use as reference (grand total) }
\item{\code{ sref_id }}{ The id for the variable to use as subreference (sub total)   }
\item{\code{ is_ref }}{ A flag to indicate if: 0-no reference variable, 1-reference, 2-subreference }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
vars
}
\description{
The list of variables with some of their characteristics
}
\keyword{datasets}
