% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars_refs}
\alias{vars_refs}
\title{vars_refs}
\format{
A data.table with the following columns:
\describe{
\item{\code{ var_id }}{ The code for the variables }
\item{\code{ level }}{ The depth of reference, with \code{C} being the \emph{cross} reference }
\item{\code{ ref_id }}{ The code of the reference variable }
}
}
\usage{
vars_refs
}
\description{
The references code for each variable and level of depth (also indent for reporting)
}
\keyword{datasets}
