% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WARD}
\alias{WARD}
\title{WARD}
\format{
A \code{sf} dataframe with only one \code{WARD} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_WARD)}.
}
\usage{
WARD
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 7,878 \emph{Electoral Ward} in England and Wales.
}
\details{
Built by dissolving the \code{OA} boundaries using the \code{lookups} table.

Last Updated: Dec-21
}
\keyword{datasets}
