% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WARDo}
\alias{WARDo}
\title{WARDo}
\format{
A \code{sf} dataframe with only one \code{WARD} column for the corresponding \emph{ONS} codes.

For further details see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=-modified&tags=all(BDY_WARD)}.
}
\usage{
WARDo
}
\description{
Digital Vector Boundaries in \code{sf} format and \emph{WGS84} CRS (\emph{EPSG} 4326) for all 7,878 \emph{Electoral Ward} in England and Wales.
}
\details{
This is the result of the conversion of the original ONS boundaries in \emph{shapefile} format
with a \emph{OSGB36} CRS (\emph{EPSG} 27700), simplified at a \emph{20\%} rate.

Last Updated: Dec-21
}
\keyword{datasets}
