% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domains}
\alias{domains}
\title{domains}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The identifier for the Domain }
\item{\code{ description }}{ The description for the Domain }
}
}
\usage{
domains
}
\description{
The list of tables with some of their characteristics
}
\keyword{datasets}
