% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_language}
\alias{dt_language}
\title{Languages and UK Proficiency}
\format{
A data.table in \emph{long} format with the following columns
\describe{
\item{\code{ zone_id }}{ The \code{ONS} identifier for the Zone, foreign key to the \code{id} column in the \code{zones} table }
\item{\code{ var_id }}{ The internal identifier for the Variable, foreign key to the \code{id} column in the \code{vars} table }
\item{\code{ value }}{ The value for the Variable in the Zone}
}

The dataset includes the following Census Tables:
\describe{
\item{\code{ TS024 }}{ Main Language (detailed) }
\item{\code{ TS025 }}{ Household Language }
\item{\code{ TS029 }}{ Proficiency in English }
\item{\code{ TS026 }}{ Multiple main Languages in households }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_language
}
\description{
Resident Population and Households by main Language(s) spoken and Proficiency in English.
}
\keyword{datasets}
