% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_religion}
\alias{dt_religion}
\title{Religions}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS030 }}{ Religion }
\item{\code{ TS031 }}{ Religion (detailed) }
\item{\code{ TS075 }}{ Multi Religion households }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_religion
}
\description{
Religions
}
\keyword{datasets}
