% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_sexage}
\alias{dt_sexage}
\title{Sex and Age}
\format{
A data.table including the following Census Tables:
\describe{
\item{\code{ TS007 }}{ Age by single year }
\item{\code{ TS007 }}{ Age by five years' classes }
\item{\code{ TS008 }}{ Sex }
\item{\code{ TS009 }}{ Sex by single year of age }
\item{\code{ TS009 }}{ Sex by five years' classes of age }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
dt_sexage
}
\description{
Sex and Age
}
\keyword{datasets}
