% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbours}
\alias{neighbours}
\title{Neighbours}
\format{
A data.table with the following columns:
\describe{
\item{\code{ type }}{ The type of the Zone as referenced in the \code{zone_types} table }
\item{\code{ idA  }}{ The ONS identifier for a Zone }
\item{\code{ idB  }}{ The ONS identifier for the first order neighbours of the Zone with code \code{idA}}
}
}
\usage{
neighbours
}
\description{
This dataset contains the \emph{1st order neighbours} for all the Zones listed in the \code{zones} table.
There are no repetitions in \code{idB} for \code{idA}, as obviously any Zone is a neighbour of itself.
}
\keyword{datasets}
