% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tables}
\alias{tables}
\title{tables}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The identifier for the Table }
\item{\code{ code }}{ The ONS code for the Table }
\item{\code{ description }}{ The descritpion for the Table }
\item{\code{ domain_id }}{ The Domain the table is included, a foreign reference to the column \code{id} in the \code{domains} table }
\item{\code{ min_type }}{ The minimum type of geographic hierarchy the data are available }
\item{\code{ units }}{ The units in whi data are expressed }
\item{\code{ url }}{ The direct url for the bulk download of data }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
tables
}
\description{
The list of tables with some of their characteristics
}
\keyword{datasets}
