% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vars}
\alias{vars}
\title{vars}
\format{
A data.table with the following columns:
\describe{
\item{\code{ id }}{ The code for the variables, built from ONS codes and file format }
\item{\code{ prog_id }}{ The \emph{internal} code for the variables, mostly used for ordering }
\item{\code{ description }}{ The description for the variables }
\item{\code{ table_id }}{ The Domain the table is included, a foreign reference to the column \code{id} in the \code{domains} table }
\item{\code{ level }}{ The depth of reference for the data. All variables with a level of \code{0} are also grouped together in the \code{main_refs} table. }
\item{\code{ loaded }}{ }
}

For further details, see the \href{https://www.nomisweb.co.uk/sources/census_2021}{Nomis website}
}
\usage{
vars
}
\description{
The list of variables with some of their characteristics
}
\keyword{datasets}
