% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zone_types}
\alias{zone_types}
\title{Zone Types}
\format{
A data.table with the following columns:
\describe{
\item{\code{type}}{ An acronym for the Geography, used as its \code{id} throughout }
\item{\code{id}}{ A numeric identifier for the Geography, used mostly for ordering }
\item{\code{ons_id}}{ The id of the Geography used by ONS }
\item{\code{name}}{ The Name for the Geography }
\item{\code{level}}{ The level of details of the Geography in terms of hierarchy }
\item{\code{max_child}}{ The \code{type} of Geography that nest exactly into the current Geography }
\item{\code{min_parent}}{ The \code{type} of Geography that all the Zones of the current Geography nest exactly into }
\item{\code{nE}}{ The total number of Zones in England }
\item{\code{nW}}{ The total number of Zones in Wales }
\item{\code{nEW}}{ The total number of Zones in England and Wales }
\item{\code{is_exact}}{  }
\item{\code{cbo_filter}}{  }
}

For further details, see \url{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(PRD_RGC)} and
\url{https://www.arcgis.com/sharing/rest/content/items/0d5d0ad97af04e18bc584e4fc3bc62de/data}
}
\usage{
zone_types
}
\description{
A list of all the Geographies included in the package(apart from the \emph{Output Areas}, listed in the \code{lookups} table)
}
\keyword{datasets}
