% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\title{Zones}
\format{
A data.table with the following columns:
\describe{
\item{\code{type}}{ The geographic hierarchy the Area belongs to, as referenced in the \code{zone_types} table }
\item{\code{id}}{ The \code{ONS} identifier for the Area }
\item{\code{name}}{ The name for the Area  }
\item{\code{parent}}{ The code of the zone in which the current zone is contained as immediate level in its most direct hierarchy}
\item{\code{country}}{ The Country the Area belongs to (either \code{E}ngland or \code{W}ales)}
\item{\code{perimeter}}{ The perimeter }
\item{\code{area}}{ The area }
\item{\code{x_lon}}{ The longitude coordinate of the geometric centroid }
\item{\code{y_lat}}{ The latitude coordinate of the geometric centroid }
\item{\code{px_lon}}{ The longitude coordinate of the pole of inaccessibility }
\item{\code{py_lat}}{ The latitude coordinate of the pole of inaccessibility }
\item{\code{wx_lon}}{ The longitude coordinate of the population weighted centroid }
\item{\code{wy_lat}}{ The latitude coordinate of the population weighted centroid }
\item{\code{bb_xmin}}{ The minimum longitude coordinate of the bounding box surrounding the polygon representing the location }
\item{\code{bb_ymin}}{ The minimum latitude coordinate of the bounding box surrounding the polygon representing the location }
\item{\code{bb_xmax}}{ The maximum longitude coordinate of the bounding box surrounding the the polygon representing the location }
\item{\code{bb_ymax}}{ The maximum latitude coordinate of the bounding box surrounding the the polygon representing the location }
}

For further details, see the \code{Names and Codes} section within \url{https://geoportal.statistics.gov.uk/}
}
\usage{
zones
}
\description{
A list of all the areas included in the \code{lookups} table, apart from the \emph{Output Areas}, together with their names and some geographic characteristics.
}
\keyword{datasets}
